added array.py in io package
