+++ larray/io/__init__.py
@@ -1,3 +1,5 @@
 from __future__ import absolute_import, division, print_function
 
 from larray.io.excel import open_excel
+from larray.io.array import *
+from larray.io.session import *
