+++ larray/io/session.py
@@ -4,9 +4,9 @@ import os
 from collections import OrderedDict
 from pandas import ExcelWriter, ExcelFile, HDFStore
 
-from larray.core.array import df_aslarray, read_csv, read_hdf
 from larray.util.misc import pickle
 from larray.io.excel import open_excel
+from larray.io.array import df_aslarray, read_csv, read_hdf
 
 try:
     import xlwings as xw
