+++ larray/tests/test_array.py
@@ -13,11 +13,11 @@ except ImportError:
     xw = None
 
 from larray.tests.common import abspath, assert_array_equal, assert_array_nan_equal
-from larray import (LArray, Axis, LGroup, union, read_hdf, read_csv, read_eurostat, read_excel, open_excel,
-                    zeros, zeros_like, ndrange, ndtest, ones, eye, diag, stack,
+from larray import (LArray, Axis, LGroup, union, zeros, zeros_like, ndrange, ndtest, ones, eye, diag, stack,
                     clip, exp, where, x, mean, isnan, round, from_lists, from_string)
 from larray.core.axis import _to_ticks, _to_key
-from larray.core.array import df_aslarray
+from larray.io.array import df_aslarray, read_hdf, read_csv, read_eurostat, read_excel
+from larray.io.excel import open_excel
 
 
 class TestValueStrings(TestCase):
